+++ geolocator_platform_interface/test/src/models/position_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: prefer_const_constructors
+
 import 'package:flutter_test/flutter_test.dart';
 import 'package:geolocator_platform_interface/geolocator_platform_interface.dart';
 
@@ -449,6 +451,44 @@ void main() {
       );
     });
 
+    test(
+        'hashCode should not match when the androidPosition property is different',
+        () {
+      // Arrange
+      final firstPosition = Position(
+        longitude: 0,
+        latitude: 0,
+        timestamp: DateTime.fromMillisecondsSinceEpoch(0),
+        accuracy: 0,
+        altitude: 0,
+        altitudeAccuracy: 0,
+        heading: 0,
+        headingAccuracy: 0,
+        speed: 0,
+        speedAccuracy: 0,
+        isMocked: true,
+      );
+      final secondPosition = Position(
+        longitude: 0,
+        latitude: 0,
+        timestamp: DateTime.fromMillisecondsSinceEpoch(0),
+        accuracy: 0,
+        altitude: 0,
+        altitudeAccuracy: 0,
+        heading: 0,
+        headingAccuracy: 0,
+        speed: 0,
+        speedAccuracy: 0,
+        isMocked: true,
+      );
+
+      // Act & Assert
+      expect(
+        firstPosition.hashCode != secondPosition.hashCode,
+        true,
+      );
+    });
+
     test('hashCode should not match when the floor property is different', () {
       // Arrange
       final firstPosition = Position(
@@ -530,7 +570,6 @@ void main() {
         speedAccuracy: 0,
         isMocked: false,
       );
-
       // Act
       final readablePosition = position.toString();